pyproject.toml
[project]
name = "rainfall-notify"
version = "0.1.0"
description = "雨の降り始めを通知するアプリです．"
authors = [
    { name = "KIMATA Tetsuya", email = "kimata@green-rabbit.net" }
]

readme = "README.md"
requires-python = ">= 3.10"

dependencies = [
    "docopt-ng>=0.9.0",
    "influxdb-client[ciso]>=1.44.0",
    "my-lib @ git+https://github.com/kimata/my-py-lib@09ac0873aefda2eba2e05d653adcf2facad328cc",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rainfall"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.rye]
managed = true
dev-dependencies = [
    "pytest-cov>=5.0.0",
    "pytest-freezegun>=0.4.2",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.14.0",
    "pytest-playwright>=0.5.1",
    "pytest-xdist>=3.6.1",
    "pytest>=8.3.2",
]

[tool.pytest.ini_options]
minversion = "6.0"

# addopts = "--verbose --log-cli-level=DEBUG --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys  --html=tests/evidence/index.htm --self-contained-html --cov=src --cov-report=html -vv"

addopts = "--verbose --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys  --html=tests/evidence/index.htm --self-contained-html --cov=src --cov-report=html -vv"


testpaths = [
    "tests",
]
filterwarnings = [
    "ignore:The hookimpl CovPlugin.pytest_configure_node uses",
    "ignore:The hookimpl CovPlugin.pytest_testnodedown uses",
    "ignore::DeprecationWarning:pytest_freezegun",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",

  # 以下，未使用関数
  # webapp/log.py
  "if sqlite is not None:",
]

[tool.coverage.html]
directory = "tests/evidence/coverage"
